requirements.txt
google-cloud-bigquery==1.22.0
